Update README.md
This code, calculates the personal inflation rate for a specified year based on the user's input for expenditure data. Here's a short description of the code:

The user is prompted to enter the year for which they want to calculate the personal interest and the number of expenditure categories they have data for.

Two variables, totalPrevExp and totalYofExp, are initialized to keep track of the total expenses for the previous year and the year of interest, respectively.

The code enters a while loop that iterates for the number of expenditure categories specified by the user. Within each iteration:

The user is prompted to enter the expenses for the previous year and the year of interest, and these values are added to the respective total expense variables.
The numExpen variable is decremented by 1 to control the loop's execution.
The inflation rate is calculated using the formula: inf_rate = ((totalYofExp - totalPrevExp) / totalYofExp) * 100.

The program prints out the calculated personal inflation rate for the specified year.

An if-elif-else block determines the type of inflation (low, moderate, high, or hyper) based on the calculated inflation rate.

Finally, the code prints out the type of inflation based on the result obtained in the previous step.